Refactor UTF-8 decoder
The old decoder was more or less written during a few hours of waiting
time together with Roberto at Budapest airport after slcon 2 in 2015.

Mattias rightfully pointed out some issues with it, namely

  - it contains unnecessary null byte handling
  - overlong encoding detection is inefficient
  - surrogate code points are not rejected
  - code points beyong U+10FFFF are not rejected

I personally had some notes on the decoder and what I didn't like,
especially its semantics. The problem with processing data is that
you can end up in a situation where you need more than is provided.
It is difficult to express these conditions with error codes, especially
when you also return a codepoint.

Another issue was error handling, that was inadequate in some cases.

Inspired by strlcat from OpenBSD, which returns a length that can be
larger than the total array capacity to indicate a truncation, I changed
the semantics of the decode-function as follows:
When there's any issue, the codepoint is set to CP_INVALID (naturally).
The length return value gives more indication for the error, though
it's not exhaustive, which doesn't matter though.

When the given array is shorter than necessary, the returned length
shows how much data _is_ needed to properly process it. In any other
case, the number of "consumed" bytes is returned (in error cases
excluding the "unexpected" byte). So, comparable to strlcat, one can
use the decode()-function safely like

   char *arr = ...; size_t arrlen = ...;
   uint32_t cp;
   size_t len;

   if ((len = grapheme_cp_decode(&cp, arr, arrlen)) > arrlen) {
	/*
	 * array too short; read in more data or just surrender and
	 * continue with cp regardless
	 */
   }

   ...

The check for CP_INVALID isn't even necessary in most cases and
checking the returned length alone is sufficient.

Signed-off-by: Laslo Hunhold <dev@frign.de>