Add UTF-8-encode-function
Merely to detect grapheme clusters, the reasoning behind adding
an encoding-function is not immediately apparent. The main reason
for it is because some decoding-scenarios actually change the text
representation (by identifying invalid codepoints and outputting
them as such).
The user should have the chance to output a "processed" stream.

A minor benefit with very little overhead is that this encoding
function is just useful in general.

Signed-off-by: Laslo Hunhold <dev@frign.de>