Expose grapheme_cp_{en,de}code() and grapheme_boundary()
After the preparation, we can now expose these three functions in
grapheme.h, as suggested by Mattias.
In this context, we get rid of the Codepoint-typedef, as there is no
need to opaquely define uint32_t. A codepoint is just a number, and thus
let's stop with the "Rune", "Codepoint", etc. naming-nonsense!

Moving everything into grapheme.h, there is also no need for boundary.h
and codepoint.h, which we reflect in the Makefile.

Signed-off-by: Laslo Hunhold <dev@frign.de>