Rename grapheme_len() to grapheme_bytelen() and refactor manual
I wasn't happy with the previous name, because it can be a bit confusing,
given there are functions like strlen() which count the number of bytes
in a string and one might assume that grapheme_len() counts the number
of grapheme clusters in a string.

Calling it grapheme_bytelen() clears this confusion up, as one can then
tell that it's about the number of bytes in a grapheme cluster and not
the number of grapheme clusters.

The manual was refactored inspired by the high standards set by the
OpenBSD manuals.

Signed-off-by: Laslo Hunhold <dev@frign.de>