Refactor grapheme_boundary()
Get rid of the macro-mess, make the code more readable with better
comments and formatting and allow the state-pointer to be NULL, which
is as if the value state is pointing to contained 0, only with the
difference that the state is not updated.

I would, actually, prefer to return some kind of error, but I like
the fact that the function can't really error out as is, which is why
I've chosen this compromise.

Signed-off-by: Laslo Hunhold <dev@frign.de>