Replace awk-scripts with C programs for data-parsing
Even though one can expect POSIX awk(1) to be present on almost all
conceivable systems, I personally must admit that I was never
comfortable with it and had to really bend it to support the features
necessary for the Unicode data table parsing (most prominently,
parsing hexadecimal numbers).

It is common to write short awk-invocations to parse line-oriented
data, but it hits its limits at the given scale. Much finer-grained
control is possible in C, with the added benefit that code-reuse is
possible and people familiar with C can now also debug the data parsing.
All in all, it adds a few lines overall, but only marginally if you
consider the fact that C is such a low-level language.

As a result, libgrapheme now only needs POSIX make(1) and a C99
compiler, while simplifying the Makefile a bit as well.

Signed-off-by: Laslo Hunhold <dev@frign.de>