Don't compile tests by default
If someone decides to call "make test", it takes only a few microseconds
to compile the tests. Given they are more of a development tool and
might only be called when explicitly desired, it's a good measure not
to "litter" the working directory with compilates, also saving time
and energy for 99% of the cases.

Parsing the tables and compiling the library takes just 100ms on my
computer, but ignoring small things like these brought us many problems
we see in computing today, as small things tend to add up.

Signed-off-by: Laslo Hunhold <dev@frign.de>