Refactor data-tables and lookup-code to be more universal
Previously, we had an explicit list of property-tables in boundary.c,
but this shouldn't be the place to put them, especially if we plan to
use the tables somewhere else, too. Instead, structure the data by
better by emitting an enum for each datatype and base the rest of the
code on it.

This turns out for boundary to require two heisenstates (see below) for
the emoji- and grapheme-break-tables, which is more than before, but
that was only possible because we reduced the generalizability of the
code.

The advantage is that if the Unicode specification adds another type
of character and implements it in the algorithm, it is as simple as
adding an element to the properties-array in the data-generator (e.g.
data/grapheme_boundary.c) which in turn automatically adds it to the
enum and you are free to access it from the code.

The specific changes are summarized below:

- Add heisenstate-struct to handle partially-known states to prevent
  multiple identical lookups. Previously, we already kept track of it,
  but handled the states by hand.
- Remove explicit table-listing in src/boundary.c and adapt code
- New datatable-format that is necessary to have an enum-indexed table
- Rename data/util.h to data/datautil.h

Signed-off-by: Laslo Hunhold <dev@frign.de>