Update to Unicode 14.0.0
There have been no changes to the grapheme cluster boundary
algorithm in UAX #29 section 3.1.1 (see [0]) and only some additions,
which partially affect the ruleset (emoji detection, new alphabets).

I thought about making the Unicode version a variable in config.mk,
but it's not always as simple as this time. To give an example, some
Unicode revisions in the past completely overhauled the grapheme
cluster boundary detection and even made it stateful.

[0]:https://unicode.org/reports/tr29/#Grapheme_Cluster_Boundary_Rules

Signed-off-by: Laslo Hunhold <dev@frign.de>