Refactor data-generation and library structure
What I always didn't like was the fact that you would have to have
two heisenstates in grapheme_boundary() (one for the grapheme-proptable
and one for the emoji-proptable). This unnecessarily complicated the
handling a little bit, even though there is still room for improvement.

A new folder gen was created to contain the generation tools. The data
folder from now on only contains data files.

Now gen/util.c contains all necessary functions to properly parse
property files (and test files) and you merely have to create an
"order list" (e.g. in gen/grapheme.c and gen/grapheme-test.c) and then
are good to go. This doesn't immensely remove code duplication, but
will come in handy in the future.

Additionally, src/boundary.c was moved into src/grapheme.c so there's
only one object file pulling in the data-table. This separation makes
the structure of the program clearer and helps the linker discard
unused library elements.

The heisenstate was increased to 64 bits for future use.

Signed-off-by: Laslo Hunhold <dev@frign.de>