Refactor API ("lg_" prefix, better naming scheme)
The "grapheme_" prefix was sadly a bit confusing so it now switches
to the "lg_" prefix which also will not get in the way too much.

"_nextbreak" and "_isbreak" as a general form makes clearer what
we actually do.

"utf8_decode" and "utf8_encode" instead of "cp_decode" and
"cp_encode" greatly improves readability and removes any doubt about
what these functions do. libgrapheme is usable with any other encoding
via the "_isbreak"-functions, but you'll have to decode yourself, but
it should be clear by now that UTF-8 should be used everywhere. :)

Signed-off-by: Laslo Hunhold <dev@frign.de>