Severely speed up grapheme cluster break detection
Instead of discarding all properties we determined for two adjacent
codepoints, we use the given state (we have to have for the flags) to
shift the properties of the right character and move them to the left
character. This change leads to a speedup of ~25%.

On the API, it changes the "int *" state argument to an opaque type
"LG_SEGMENTATION_STATE *". Any normal library would just define an
internal struct and export a function "init_state" or something and
be done with it, but I didn't like the idea of having heap-allocations
in libgrapheme and wanted to make it possible to just do

   LG_SEGMENTATION_STATE state;
   lg_grapheme_isbreak(a, b, &state);

but have also the choice to do

   LG_SEGMENTATION_STATE *state;
   state = malloc(sizeof(*state));
   lg_grapheme_isbreak(a, b, &state);

whatever is preferred or makes sense in the context.

Doing this is pretty difficult, because allowing C to know the storage
size of a type through the API is only possible if it knows the
implementation details. My first attempt was to define a struct with
identical structure in grapheme.h (only with "anonymous" structs
rather than the internal heisenstate-structs), but after consulting
the standard this is undefined behaviour and breaks strict aliasing
rules. So I bit the bullet and gave the definition in the header,
indicating with "internal" that these types should not be used.

In most cases, users will probably never get in contact with this data
structure, as lg_grapheme_nextbreak() is still as clean as before.

Signed-off-by: Laslo Hunhold <dev@frign.de>