Refine types (uint8_t -> char, uint32_t -> uint_least32_t)
The type uint32_t is not guaranteed by the standard to be present,
but it guarantees uint_least32_t. If a libgrapheme-user passes a pointer
to an uint32_t (instead of uint_least32_t) there will be no problem,
as the presence of uint32_t immediately implies uint32_t ==
uint_least32_t. However, we won't depend on it internally and are
strict with using uint_least32_t. The type name is a mouthful, but still
clearer and not much longer than "long int" (which is guaranteed to be
at least 32 bits).

Regarding uint8_t, it was a bit clumsy to require it in the API. C does
not guarantee that a byte is actually an octet (i.e. char can have
more than 8 bits), and even though the relevance of non-8-bit-char
seems to be waning, I don't want to rely on that. But more importantly,
accepting "char *" saves some casts on the user-side.
Adapting the lg_utf8_* functions is trivial, as it requires just
being careful with casts. The cast "signed char" <-> "unsigned char"
is unproblematic, so every time we need the bit representation, we
explicitly cast to unsigned char and are done with it. Likewise, every
time we write to a char, we make sure that what we pass is explicitly
an unsigned char.
This became a bit awkward in the test cases where we have char-arrays
with hex literals. As C does not really have a concept of a sub-int
literal, all hexadecimal literals had to first be explicitly cast to
unsigned char, but that's it.

One more aspect where we've become more portable. :)

Signed-off-by: Laslo Hunhold <dev@frign.de>