Assume nothing, assert what we want in the API
After a long discussion with and a lot of input from Michael Forney,
which I am very thankful for, we came to the conclusion that accepting
"char *" is problematic and implies a lot of things that reach back
decades. In fact, "char" is so broken that the functions defined in
bloody "string.h" operate on "void *". Go figure...

It would have been possible to simply change the API to work with
"unsigned char *" to avoid signed integer representation shenanigans,
but at that point, if the user cannot use string literals and char-arrays
without a cast anyway, we might as well go all the way and be clear
about what we really want (array of octets). Supporting char doesn't
really make sense when char is not 8-bit, given then the user would
have to do a lot of hand-crafting to make it work anyway. Even if
such systems with chars >8 bit were common (which they aren't at all), I
would still have probably made this decision here.

UTF-8 is an 8-bit encoding and it only makes sense on octet-arrays.
This might even motivate libgrapheme-users to be explicit about it
in their code as well. Otherwise, given POSIX guarantees that char
is exactly 8 bits, simply casting to (uint8_t *) will work.

C's type system is already weak enough, so where possible ambiguities
like that should be resolved in favor of more robust solutions in my
opinion.

Signed-off-by: Laslo Hunhold <dev@frign.de>