Change lg_grapheme_isbreak() return type from int to bool
Just as a disclaimer, I still really like int as a return type for
functions to indicate errors (return 0 means success and anything >0
means failure). For a library though one would most certainly use
an enum to name those errors, too, which would then turn the function
signature to "enum foo_errors foo(...)".

lg_grapheme_isbreak() cannot fail, however, and as in the previous
commit on reducing implicit assumptions on char, reducing assumptions
and making clearer what a return value can entail is good. This is
why the return type is changed to bool.

The bool type is C99 and even the more "independent" compilers like scc
support it. If it is internally expanded to an int, so be it, but it
helps readability.

Signed-off-by: Laslo Hunhold <dev@frign.de>