Optimize bsearch-comparison-function for cp-ranges
Instead of having two comparisons all the time, reduce it to one with
a clever case-shuffling. Also, instead of calculating the "offset"
of the cp from the range, simply indicate by sign if we're above or
below (the bsearch() function/binary search algorithm cannot make use
of the magnitude anyway), which also conveniently avoids possible
overflows.

This change leads to a speedup of ~6%.

Signed-off-by: Laslo Hunhold <dev@frign.de>