Make lists in Makefile more idiomatic and avoid breaks
Thanks Quentin Rameau for the remark regarding the more idiomatic
way to specify lists where also the last element has an explicit
line-break followed by an empty line.

Also avoid breaks later on in the code: This breaks the line-length
a bit, but has an effect on the output. I prefer a cleaner output
(also in the build-logs) over one overlong line.

Signed-off-by: Laslo Hunhold <dev@frign.de>