Improve a small edge-case in lg_utf8_decode()
Okay, this case is really crazy but possible: Before this change,
when we encountered e.g. a 0xF0 (which indicates a 4-byte-UTF-8
sequence and implies 3 subsequent continuation bytes) but have a
string-length of e.g. 2, we would automatically return 4 (> 2) no matter
how the following bytes look like to indicate that we need a larger
buffer.

However, it's actually necessary to check the subsequent bytes until
the buffer-end as we might have a case like

   0xF0 0x80 0x00

where 0xF0 is followed by a single continuation byte but then the
continuation stops and we have a NUL-byte. It's more expected to
return 2 in such a situation because we obtain more information about
the string by inspecting the continuation bytes instead of throwing
our hands up so early.

Also add this to the test-cases of the decoder to prevent any
regressions.

Signed-off-by: Laslo Hunhold <dev@frign.de>