Rename functions/defines/files from "grapheme" to "character"
It was always confusing to have "grapheme" used in different contexts.
One is the library name, which is also the prefix for all constants,
the other is to indicate we have functions for analyzing grapheme
clusters.

We rename all functions related to graphemes to operate on "characters"
instead, where these are user-perceived characters. This naming choice
is not out of the ordinary and also what libunistring for instance uses.

Additionally, rename gen/grapheme.c to gen/character-prop.c to indicate
we extract properties, improving readability. This also removes a bit
of ambiguity regarding internal constants prefixed with GRAPHEME_, which
might suggest that these were "officially" from grapheme.h, even though
they serve only an internal use for characters specifically.

Signed-off-by: Laslo Hunhold <dev@frign.de>