Rename GRAPHEME_SEGMENTATION_STATE to GRAPHEME_STATE
The name was getting a bit long and the user doesn't really have to
care about the fact it's specifically a segmentation state. If we,
for any reason, require another state type, we could simply define
a union accordingly:

typedef union grapheme_internal_state {
	struct grapheme_internal_segmentation_state {
		...
	} segmentation;
	struct grapheme_internal_whatever_state {
		...
	} whatever;
} GRAPHEME_STATE;

This would not change the outside behaviour (given we defined
GRAPHEME_STATE as an opaque state), but we guarantee that

   GRAPHEME_STATE state = { 0 };

and

   memset(&state, 0, sizeof(state));

continue to work.

Signed-off-by: Laslo Hunhold <dev@frign.de>