Rename API functions to improve readability
I thought about how to address the fact that "isbreak" and "nextbreak"
kind of breaks the snake case, but "grapheme_character_is_break" sounds
convoluted.

The solution is to loosen the naming a bit and not require the
"component" (in this case "character") to immediately follow the
"grapheme_" prefix. Instead, the "is" and "next" keywords are brought
to the front, which improves the readability substantially and the
functions are well-grouped into "is" and "next" functions.

Analogously, it makes more sense to "decode_utf8" than "utf8_decode",
so this was changed as well, including going back to
GRAPHEME_INVALID_CODEPOINT, which just rolls off the tongue better.

Signed-off-by: Laslo Hunhold <dev@frign.de>