Add generator for compressed O(1)-lookup-table
Up to this point, the idea behind property-checks was to have a sorted
list and run something like binary search on it. Even though binary
search works really well and is O(log(n)), with our multiple tables
we suffer from the effect that

   sum_{i=1..N} O(log(i)) > O(log(sum_{i=1..N} i))

In simpler terms, the asymptotic gains are largest if you have one
large sorted list that you run binary search on. So the first thing I
tried out was putting all into one long range-list and add the break
property to each array.

While this worked and was much faster, it still was not satisfactory.

Preliminary experiments showed that it is possible to go with a really
crazy approach: Actually have a mapping of each of the 0x110000
codepoints separated into two separate tables, where the compression
works such that the first table indexes into the highest 2 bytes
(shifted 1 byte to the right) and then you have a subtable indexing
into the lowest byte.
The main advantage is that you can compress it by pointing to identical
sub-sequences. This might sound very rudimentary, but it gives a
compression ratio of 97.91%, which means that the lookup-table is only
around 55K, which is totally acceptable.

Really crazy are the speed-gains, because you end up with a branchless
lookup. utf8proc employs a similar method using hash-tables during
processing, but it has a bug that makes the compression around 20% worse
than what we have here. Additionally, their generator is written in
Ruby.

Stay tuned for the subsequent additions in the grapheme break detection
algorithm. I'm not exaggerating when saying that libgrapheme will by far
be the fastest implementation out there, but the code needs a bit of
cleanup I won't be able to finish today.

Until then I wish you a happy new year!

Signed-off-by: Laslo Hunhold <dev@frign.de>