Create separate comparative benchmark/ folder
Benchmarks shouldn't really be part of the functional test and are
only meaningful in direct comparison with other implementations, which
obviously should not be a install-time-dependency. Instead, benchmarks
are meant as a development-tool, and thus deserve their own separate
folder.

So, which library should we compare ourselves to? There are many
options, like libunistring, libicu and libutf8proc, but at the end only
the latter remains, for the following reasons: libunistring's
implementation is incorrect, as it is not stateful and thus probably
fell victim to the semantic change of the standard a few years ago,
which makes any benchmark against it meaningless. libicu does not offer
a way to simply check two code points with a manual state. libutf8proc
is a great comparison, because it's the backbone of the modern Julia
language and received a lot of performance optimization.

Measuring ourselves against it makes a lot of sense.

Regarding the implementation: Given we don't use the return value of the
break check, I added function-level-optimization attributes to the
function. However, they are behind the feature-check "__has_attribute"
ensuring maximum portability even for compilers which don't offer that.

Signed-off-by: Laslo Hunhold <dev@frign.de>