Introduce mostly branchless character break detection
Given the ever-increasing CPU-cache-sizes compared to the attempts of
branch-prediction, where the CPU executes from cache-miss to cache-miss,
the optimum is to have basically no branches at all.

In the case of character-break-detection, this would necessitate a
lookup-table over all 0x110000 (~1.1 million) codepoints that would
be, if we calculate one byte per entry, around 1.1MB, which is totally
excessive.
However, the generator added in 759f02fbe74f11f64ba451b3a18090cbbb10add2
achieves a compression ratio of 97.91% using a two-stage-compression,
yielding a lookup-table of size roughly around 55KB.

55KB easily fits even within L1 (which is often around 64KB), let alone
L2 and L3 which are getting even larger. The CPU does not have to do any
branch prediction as it can anticipate the cache-access.

More interesting are the break rule: Different from libutf8proc, which
also employs the two-stage approach to determine character-properties,
it then uses normal branching to actually apply the algorithm.
However, we can do it better: Given the limited number of break
properties, we can formulate lookup-tables to determine breaks,
optionally even including flags which can be added as offsets within the
tables. The flag-updates themselves can also be achieved with LUTs.

So in the end, we don't branch, as the CPU simply accesses
easily-predictable cache-areas. The branches still left can be evaluated
later on for further optimization.

The benchmarks speak for themselves: We are 50-150% faster than
libutf8proc now, yielding a total speedup of around 40x.

Signed-off-by: Laslo Hunhold <dev@frign.de>