Generate separate utf8proc_int32_t buffer to preserve strict aliasing
This clearly shows why it's never a good idea to roll your own types
and to better stick with ones provided by the standard library.

Even if the custom type in libutf8proc was defined as an unsigned
32-bit integer type, it could be changed at any point (e.g. to
uint_fast32_t which might default to an unsigned 64 bit type). So we
can't simply cast between the pointers anyway, even if we didn't care
about strict aliasing.

Signed-off-by: Laslo Hunhold <dev@frign.de>