Mark likely branches
The likely() and unlikely() macros (known from the Linux kernel) in
src/util.h are defined in a portable way.

This addition brings a performance improvment of around 6%, which also
shows how well-optimized grapheme_is_character_break() already is. One
break check for two codepoints now takes only around 10µs on average on
my machine, which is insanely fast when you consider the complexity of
the algorithm behind it.

Signed-off-by: Laslo Hunhold <dev@frign.de>