Calculate benchmark-times as average time taken for each call
Add a unit-parameter and calculate all times as average time to call a
function once. Previously focusing all on the relative performance, I
totally overlooked that the printed times were not for a single
comparison but one iteration of the input buffer.

In this case, one codepoint break detection on my machine only takes
around 7ns!!! This means that the average processing speed is, if we
assume on average 1.2-1.5 bytes per codepoint, around 200 MB/s.

To put this into perspective, a terminal buffer may have a grid of
70x200 = 14000 visible characters. Calculating breakpoints for all
of them would only take around 0.1ms, which is, even if it was done on
every draw call, a tiny fraction of the 18ms needed to draw at 60FPS.

Signed-off-by: Laslo Hunhold <dev@frign.de>