Rename gen/properties.c to gen/character-properties.c
At first I thought we would have one big lookup-table for all
properties, but it doesn't make sense for multiple reasons. Even though
compression remains great even for more than one property (it goes down
to roughly 95-96%), this still has a drastic impact on performance.
Additionally, as mentioned earlier, this allows better LTO for the use
of a single or few functions (which should be the common case), because
each data-table is separate.

Signed-off-by: Laslo Hunhold <dev@frign.de>