Properly handle cp == NULL in grapheme_decode_utf8()
During refactoring I totally forgot about it. Instead of adding a
check every time we do anything, we save a lot of branching by doing
a single branch in the beginning, optionally setting cp, if NULL, to a
pointer to a local dummy variable.

Now it works as expected and documented, given my goal is that there
should be no case where a function segfaults due to a passed NULL
pointer.

Thanks a lot to Hécate (retro-freedom.nz) for reporting this!

Signed-off-by: Laslo Hunhold <dev@frign.de>