Implement line-segmentation
This was definitely the most difficult of all the
segmentation-algorithms, but it worked out in the end. The concepts
worked out in the other algorithms helped a lot.

The central idea was to have the fastest LUTs possible, which is why
it was implemented like that. For the line classes it required some
postprocessing and a few tricks, but the LUT in the end turns out
to be roughly 40K, which is very small.

It was interesting to see that the Unicode specifies the line breaking
algorithm as one that could be implemented like the grapheme-break-
algorithm (i.e. pairwise check with a state), but one single rule
prevents this. I didn't catch this at first, as this rule is part of
a "custom tailoring" so heavily recommended by the spec that the
test-cases rely on it. To be fair, it yields much better
number-segmentation.

The code could use some refactoring, but the primary goal was to pass
the over 7600 line-break-tests. Any regressions due to refactoring can
then be easily caught.

With line segmentation implemented libgrapheme has become quite useful
for a number of things previously only provided by extremely bloated
or incorrect libraries. Compiling on my system takes only a second and
the tarball is only 332K (including the data-files!). Using bzip2,
the size goes down to 180K, which is nothing considering what it does.

The final library .so/.a is around 170K, but if you statically link
against libgrapheme (with LTO), it will only pull in the LUTs it needs,
so for basic grapheme handling your binary will not grow more than 35K.

The next goal will be case mappings/detections, which should be much
simpler in term of the algorithmic challenge, but might be a bit more
difficult in terms of data-parsing.

Signed-off-by: Laslo Hunhold <dev@frign.de>