Implement case-conversion and -check
I found no implementation of Unicode case-conversion and -checking apart
from ICU (which sucks and is extremely bloated) that does it "correctly"
in the sense that not only 1:1-mappings and context-dependent
conversions (e.g. Final_Sigma) are supported.

It shouldn't be too complicated to work with this stuff, and I think
a lot of the unwillingness to properly support Unicode in most open source
projects is the fact that the APIs are horrible.

The data-parsing makes use of the major-minor-compression-scheme and
applies it to the case-casts (stored as offsets, which is novel).
Additionally, one requires a few properties (cased, case-ignorable,
uncased) as well. The entire dataset is compressed down to ~170K, which
is very very little considering what it does. No LUT exceeds 64K, so
they all still fit into the L1-cache.

Both single- and multi-mappings are supported, of course, and the
special case of a sigma occuring at the end of a word is also handled:

  'ΤΟ ΣΏΜΑ ΜΟΥ ΕΊΝΑΙ Ο ΝΑΌΣ ΜΟΥ' -> 'το σώμα μου είναι ο ναός μου',

to give an example. This way, the case-conversion and case-checking is
fully Unicode-compliant.

Besides upper- and lowercase, titlecase is also implemented. In total,
the implementation is also extremely fast, of course, at around 1.69
gigacodepoints/s (noice) on my machine.

As a sidenote, this commit also removes the data/LICENSE as a dependency
of each data-file, as it lead to the problem that it would always
pull in the files on each build. Instead, it is turned into a dependency
of all so it is ensured it is always present.

The final libgrapheme.a/libgrapheme.so weighs in at ~350K, which is
really nothing considering decompression is O(1) with a very small
multiplicative constant. Static linking will discard unused parts
anyway.

Signed-off-by: Laslo Hunhold <dev@frign.de>