Explicitly use object-files in library-generation
The macro $? only evaluates to the list of prerequisites that are
newer than the current target. While this incidentally works when
simply compiling, this leads to the library-files only including
the changed .o-file (and nothing else!) when a single source file
has been changed.

This is easily fixed by explicitly listing the object-files.

Signed-off-by: Laslo Hunhold <dev@frign.de>