Use ar(1)'s s-flag instead of invoking ranlib(1)
Tom Schwindl kindly pointed out that ranlib(1) is legacy and not even
part of POSIX anymore, given ar(1) can do the same job with the s-flag
(which is an XSI-extension, but whatever).

I'm certain there is no case where a system does not have a convenient
ranlib(1)-wrapper-script, but it's always good to follow best practices.

Signed-off-by: Laslo Hunhold <dev@frign.de>