Add bidirectional data-parser and function prototypes with structure
The bidirectional algorithm is very interesting and currently mostly
used in the context of the GNU-fribidi-library. Despite its complexity
it should be interesting trying to implement it in libgrapheme.

This commit adds the data-parsing and the first part of the algorithm
(namely paragraph level detection). Development will happen in a
separate branch to allow stabilization on the master branch in
preparation for the release of version 2.

Signed-off-by: Laslo Hunhold <dev@frign.de>