Revert "Use ar(1)'s s-flag instead of invoking ranlib(1)"
After further discussion with Roberto on the mailing list, it makes
more sense to keep ranlib(1). There are good reasons why POSIX does not
define it, given it relies on the binary format. Specifying the s-flag
for ar(1) is thus very surprising, but should not be a reason not to use
ranlib(1).

This reverts commit 5fc87aedad86a8410a360ba718096e613ddf11f6.