Separate CC into CC and BUILD_CC, add option to disable ldconfig(1)-call
The libgrapheme-Gentoo-ebuild[0] has to resort to modify the Makefile
itself given we unconditionally call $(CC) both for the tools we
compile at build-time and the final library objects.

It might however be the case that we're cross-compiling to another
architecture, where we have to make a distinction between local and
target compilers. Using BUILD_CC is an established approach, while also
having separate BUILD_*-flags for those compiler/linker-calls
respectively.

Additionally, use an LDCONFIG-variable that can be unset to not call
ldconfig after installing and uninstalling. This is necessary for
porting given e.g. Gentoo prefers to call ldconfig itself when it
deems it appropriate.

This commit also removes the useless print-target that got left over
from a long-forgotten debugging-session.

As a minor thing, we add the soname-flag to the default linker-options
to properly set SONAME for libgrapheme.so. We naturally don't do symbol
versioning, given even though there are some name-changes from version
1 the names should be stable now and new ones only added, not removed
or changed.

This should, in general, greatly simplify packaging of libgrapheme
as soon as the next version 2 is released.

[0]:https://gitweb.gentoo.org/repo/gentoo.git/tree/dev-libs/libgrapheme/libgrapheme-1.ebuild

Signed-off-by: Laslo Hunhold <dev@frign.de>